[File functionality] (Work with files in lessons` context)
- Now user cann perform all file functionality in context of lessons. User can do it if he was enrolled in course that lesson belongs.
- If user want to upload file twice then old file will be replaced.

+++ src/main/resources/db/migration/V12__create_table_homework_files.sql
@@ -3,7 +3,7 @@ CREATE TABLE `learning_courses`.`homework_files`
     `id`                BIGINT AUTO_INCREMENT UNIQUE NOT NULL,
     `user_to_lesson_id` BIGINT                       NOT NULL,
     `title`             VARCHAR(256)                 NOT NULL,
-    `size`              INT                          NOT NULL DEFAULT 0,
+    `size`              BIGINT                       NOT NULL DEFAULT 0,
 
     PRIMARY KEY (`id`),
